move : 파일 분리
작업 완료
함수별 파일 분리

작업 예정
설정창
캐릭터 순서 변경
- 순서 변경후 저장 방식 정하기
- 페이지를 나갈때 자동저장
    - 뒤로가기 버튼(navigation bar)를 두번 클릭하면 뒤로가면서 저장
    - TopBar 뒤로가기 버튼을 눌러 저장
- 저장 버튼을 눌러 저장
    - TopBar action부분에 버튼을 눌러 저장
    - BottomBar에 저장, 취소 버튼을 추가해 저장 및 취소
    - 뒤로가기 및 나가기 버튼을 누르면 변경사항 저장 안됨

업데이트 확인

로딩창
로딩 로티 만들어보기

검색
즐겨찾기 또는 검색기록 만들어 반복검색 하지 않기

각인
아르데타인의 기술은 effects 값이 없음

캐릭터 상세
viewModel에서 사용되는 공통 함수 model로 빼기

장비
빈값에 빈값인 이미지보여주기
-> list에서 find로 찾고 없다면 빈값, 있다면 채우기 이런식?
(나중에) 팔찌 상중하 옵션별로 색부여

스킬
트라이포드 이름이 길어지면 다음줄로 넘어가는데
TextChip이랑 Spacer, Text가 같이 넘어가게 하기
상세 배경이 검정 -> 고치기
