+++ app/src/main/java/com/hongmyeoun/goldcalc/view/setting/settingPage/SettingTopBar.kt
@@ -1,0 +1,65 @@
+package com.hongmyeoun.goldcalc.view.setting.settingPage
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.width
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.KeyboardArrowLeft
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavHostController
+import com.hongmyeoun.goldcalc.ui.theme.LightGrayBG
+import com.hongmyeoun.goldcalc.view.characterDetail.titleTextStyle
+
+@Composable
+fun SettingTopBar(
+    navController: NavHostController
+) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .background(LightGrayBG),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        IconButton(
+            modifier = Modifier.weight(0.5f),
+            onClick = {
+                navController.navigate("Main") {
+                    popUpTo("Setting") {
+                        inclusive = true
+                    }
+                }
+            }
+        ) {
+            Icon(
+                imageVector = Icons.Default.KeyboardArrowLeft,
+                contentDescription = "뒤로",
+                tint = Color.White
+            )
+        }
+
+        Text(
+            modifier = Modifier
+                .fillMaxWidth()
+                .weight(3f),
+            text = "설정",
+            style = titleTextStyle(),
+            textAlign = TextAlign.Center
+        )
+
+        Spacer(
+            modifier = Modifier
+                .width(32.dp)
+                .weight(0.5f)
+        )
+    }
+}
